Added a sending a reply to assignment
This commits calculates the message and sends the reply to the server
using the calcprotocol struct. The calcprotocol should be filled with
type equals to 2. After sending the message read the new message
sent from server and check if it sent 1 or 2. If one that's a success print "OK" else
print "NOT OK".